+++ examples/multimodal/components/video_encode_worker.py
@@ -15,8 +15,6 @@
 
 import argparse
 import asyncio
-import base64
-import binascii
 import logging
 import os
 import signal
@@ -24,10 +22,8 @@ import sys
 from io import BytesIO
 from queue import Queue
 from typing import AsyncIterator, Optional, Tuple
-from urllib.parse import urlparse
 
 import av
-import httpx
 import numpy as np
 import torch
 import uvloop
@@ -41,6 +37,7 @@ from dynamo.runtime.logging import configure_dynamo_logging
 sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), ".."))
 from utils.args import Config, base_parse_args, parse_endpoint
 from utils.protocol import MyRequestOutput, vLLMMultimodalRequest
+from utils.video_encode_utils import load_video_content, read_video_pyav
 from utils.video_processor import (
     calculate_frame_sampling_indices,
     get_video_metadata,
@@ -88,181 +85,10 @@ class VllmEncodeWorker:
         self._video_content_cache: dict[str, BytesIO] = {}
         self._cache_queue: Queue[str] = Queue(maxsize=CACHE_SIZE_MAXIMUM)
 
-        self._http_client: Optional[httpx.AsyncClient] = None
         self._http_timeout = 60.0
 
     def cleanup(self):
         pass
-
-    async def _read_video_pyav(
-        self, container: av.container.InputContainer, indices: np.ndarray
-    ) -> np.ndarray:
-        """
-        Decode the video with PyAV decoder. Async wrapper.
-        """
-
-        def blocking_decode():
-            container.seek(0)  # Reset container for decoding
-            processed_indices = set(indices)
-
-            # Determine min/max index to optimize decoding loop slightly
-            min_idx = 0
-            max_idx = -1
-            if len(indices) > 0:
-                min_idx = np.min(indices)
-                max_idx = np.max(indices)
-
-            if (
-                not processed_indices
-                and container.streams.video
-                and container.streams.video[0].frames > 0
-            ):
-                logger.warning(
-                    "_read_video_pyav called with empty indices for a non-empty video, attempting to read first frame."
-                )
-                try:
-                    frame = next(container.decode(video=0))
-                    return np.stack([frame.to_ndarray(format="rgb24")])
-                except StopIteration:
-                    logger.error(
-                        "Failed to read even the first frame despite non-empty indices check."
-                    )
-                    return np.array([])
-
-            decoded_frames_list = []
-            for i, frame in enumerate(container.decode(video=0)):
-                if i > max_idx and max_idx != -1:  # max_idx is -1 if indices is empty
-                    break
-                if i >= min_idx and i in processed_indices:
-                    decoded_frames_list.append(frame)
-
-            if not decoded_frames_list and len(processed_indices) > 0:
-                actual_decoded_count = 0
-                try:
-                    container.seek(0)  # Reset for counting
-                    for _ in container.decode(video=0):
-                        actual_decoded_count += 1
-                except Exception:  # Handle cases where re-decoding/counting fails
-                    pass  # Keep original error message
-                raise ValueError(
-                    f"Could not decode any frames for the given indices: {indices.tolist()}. "
-                    f"Video might be shorter than expected or indices out of bounds. "
-                    f"Actual decodable frames in container (approx): {actual_decoded_count}."
-                )
-
-            return (
-                np.stack([x.to_ndarray(format="rgb24") for x in decoded_frames_list])
-                if decoded_frames_list
-                else np.array([])
-            )
-
-        return await asyncio.to_thread(blocking_decode)
-
-    async def _load_video_content(self, video_url: str) -> BytesIO:
-        parsed_url = urlparse(video_url)
-        video_url_lower = video_url.lower()
-
-        if parsed_url.scheme in ("http", "https"):
-            if video_url_lower in self._video_content_cache:
-                logger.info(f"Video content found in cache for URL: {video_url}")
-                cached_content = self._video_content_cache[video_url_lower]
-                cached_content.seek(0)
-                return cached_content
-
-        try:
-            video_data: BytesIO
-            if parsed_url.scheme == "data":
-                if not parsed_url.path.startswith(
-                    ("video/", "application/octet-stream")
-                ):
-                    raise ValueError("Data URL must be a video type or octet-stream")
-
-                media_type_and_data = parsed_url.path.split(",", 1)
-                if len(media_type_and_data) != 2:
-                    raise ValueError("Invalid Data URL format: missing comma separator")
-
-                media_type, data_segment = media_type_and_data
-                if ";base64" not in media_type:
-                    raise ValueError("Video Data URL currently must be base64 encoded")
-
-                try:
-                    video_bytes = base64.b64decode(data_segment)
-                    video_data = BytesIO(video_bytes)
-                except binascii.Error as e:
-                    raise ValueError(
-                        f"Invalid base64 encoding for video data: {e}"
-                    ) from e
-
-            elif parsed_url.scheme in ("http", "https"):
-                if not self._http_client:
-                    await self._init_http_client()
-                    if not self._http_client:  # Double check after initialization
-                        raise RuntimeError("Failed to initialize HTTP client")
-
-                logger.info(f"Downloading video from URL: {video_url}")
-                response = await self._http_client.get(
-                    video_url, timeout=self._http_timeout
-                )
-                response.raise_for_status()
-
-                if not response.content:
-                    raise ValueError(
-                        f"Empty response content from video URL: {video_url}"
-                    )
-                video_data = BytesIO(response.content)
-                video_data.seek(0)
-                logger.info(
-                    f"Video downloaded from {video_url}, size: {len(response.content)} bytes."
-                )
-
-            elif parsed_url.scheme == "file" or not parsed_url.scheme:
-                file_path = parsed_url.path if parsed_url.scheme else video_url
-                # Ensure path is absolute or resolve relative to a known base if necessary
-                # For simplicity, assuming it's an accessible path.
-                if not os.path.exists(file_path):
-                    raise FileNotFoundError(f"Error reading file: {file_path}")
-
-                with open(file_path, "rb") as f:
-                    video_bytes = f.read()
-                video_data = BytesIO(video_bytes)
-            else:
-                raise ValueError(
-                    f"Unsupported video source scheme: {parsed_url.scheme} for URL {video_url}"
-                )
-
-            if parsed_url.scheme in (
-                "http",
-                "https",
-            ):  # Cache successfully downloaded content
-                if self._cache_queue.full():
-                    oldest_url = self._cache_queue.get_nowait()
-                    if oldest_url in self._video_content_cache:
-                        del self._video_content_cache[oldest_url]
-
-                # Store the BytesIO object directly; it will be seek(0)'d when retrieved
-                self._video_content_cache[video_url_lower] = video_data
-                self._cache_queue.put(video_url_lower)
-
-            return video_data
-
-        except httpx.HTTPStatusError as e:
-            logger.error(
-                f"HTTP error {e.response.status_code} loading video {video_url}: {e.response.text[:200]}"
-            )
-            raise ValueError(
-                f"Failed to download video {video_url}: HTTP {e.response.status_code}"
-            ) from e
-        except httpx.RequestError as e:
-            logger.error(f"Request error loading video {video_url}: {e}")
-            raise ValueError(f"Network request failed for video {video_url}") from e
-        except FileNotFoundError as e:
-            logger.error(f"File error loading video {video_url}: {e}")
-            raise
-        except Exception as e:
-            logger.error(
-                f"Error loading video content from {video_url}: {type(e).__name__} - {e}"
-            )
-            raise ValueError(f"Failed to load video content: {e}") from e
 
     async def generate(
         self, request: vLLMMultimodalRequest
@@ -284,7 +110,12 @@ class VllmEncodeWorker:
         container: Optional[av.container.InputContainer] = None
 
         try:
-            video_content_stream = await self._load_video_content(video_url)
+            video_content_stream = await load_video_content(
+                video_url,
+                self._video_content_cache,
+                self._cache_queue,
+                self._http_timeout,
+            )
 
             # Open video container using utility function
             container = await open_video_container(video_content_stream, video_url)
@@ -305,7 +136,7 @@ class VllmEncodeWorker:
                 raise ValueError(f"Container is None for {video_url}")
 
             # Decode video frames
-            clip_np: np.ndarray = await self._read_video_pyav(container, indices)
+            clip_np: np.ndarray = await read_video_pyav(container, indices)
 
             if clip_np.size == 0:
                 raise ValueError(
@@ -374,22 +205,12 @@ class VllmEncodeWorker:
             if container:
                 await asyncio.to_thread(container.close)
 
-    async def _init_http_client(self):
-        if (
-            not self._http_client or self._http_client.is_closed
-        ):  # Check if closed as well
-            self._http_client = httpx.AsyncClient(
-                timeout=self._http_timeout, follow_redirects=True
-            )
-            logger.info("HTTP client (re)initialized.")
-
     async def async_init(self, runtime: DistributedRuntime):
         logger.info("Startup started.")
         # Create and initialize a dynamo connector for this worker.
         # We'll needs this to move data between this worker and remote workers efficiently.
         self._connector = connect.Connector()
         await self._connector.initialize()
-        await self._init_http_client()
 
         logger.info("Startup completed.")
