+++ examples/multimodal/components/worker.py
@@ -247,10 +247,10 @@ class VllmPDWorker(VllmBaseWorker):
 
         if "video" in self.engine_args.model.lower():
             self.EMBEDDINGS_DTYPE = torch.uint8
-            self.EMBEDDINGS_DEVICE = "cpu"
         else:
             self.EMBEDDINGS_DTYPE = torch.float16
-            self.EMBEDDINGS_DEVICE = "cpu"
+
+        self.EMBEDDINGS_DEVICE = "cpu"
 
         # Create and initialize a dynamo connector for this worker.
         # We'll needs this to move data between this worker and remote workers efficiently.
@@ -282,7 +282,10 @@ class VllmPDWorker(VllmBaseWorker):
         )
         descriptor = connect.Descriptor(embeddings)
 
-        if request.multimodal_input.image_url is None and request.multimodal_input.video_url is None:
+        if (
+            request.multimodal_input.image_url is None
+            and request.multimodal_input.video_url is None
+        ):
             if descriptor is None:
                 raise RuntimeError(
                     "Descriptor is None in PD worker - cannot process embeddings"
@@ -309,7 +312,9 @@ class VllmPDWorker(VllmBaseWorker):
         else:
             # Use PIL image instead of image embeddings
             multi_modal_data = {
-                "image": await self.image_loader.load_image(request.multimodal_input.image_url)
+                "image": await self.image_loader.load_image(
+                    request.multimodal_input.image_url
+                )
             }
 
         # Remove the image features from the request as they are not required
