+++ examples/multimodal/utils/model.py
@@ -14,10 +14,10 @@
 # limitations under the License.
 
 import logging
-from typing import Any, Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Optional
 
 import torch
-from transformers import AutoConfig, AutoModel
+from transformers import AutoModel
 
 logger = logging.getLogger(__name__)
 
@@ -40,33 +40,6 @@ def load_vision_model(model_id: str) -> torch.nn.Module:
     return model
 
 
-def get_vision_embeddings_info(
-    model_id: str,
-) -> Tuple[Tuple[int, int, int], torch.dtype]:
-    """Calculate vision embeddings size and dtype using model config
-    Returns a tuple of (batch_size, seq_len, hidden_dim), dtype.
-    """
-    config = AutoConfig.from_pretrained(model_id, trust_remote_code=True)
-
-    if model_id == SupportedModels.LLAVA_1_5_7B:
-        seq_len = 577
-    elif model_id == SupportedModels.QWEN_2_5_VL_7B:
-        seq_len = 345
-    else:
-        seq_len = 0
-
-    if not hasattr(config, "torch_dtype"):
-        raise ValueError("Model config missing required 'torch_dtype' attribute")
-    if not hasattr(config, "hidden_size"):
-        logger.warning(
-            "Model config missing required 'hidden_size' attribute, using 4096"
-        )
-        hidden_size = 4096
-    else:
-        hidden_size = config.hidden_size
-    return (1, seq_len, hidden_size), config.torch_dtype
-
-
 def construct_mm_data(
     model: str,
     embeddings_dtype: torch.dtype,
@@ -75,25 +48,39 @@ def construct_mm_data(
     image_grid_thw: Optional[List[Any]] = None,
 ) -> Dict[str, torch.Tensor | Dict[str, Any]]:
     """Construct multimodal data for a vLLM request for models that require additional parameters alongside the embeddings"""
-    image_embeds = image_embeds.to(embeddings_dtype)
-    if model == SupportedModels.QWEN_2_5_VL_7B:
-        if image_grid_thw is not None and len(image_grid_thw) > 0:
-            grid_thw_tensor = torch.tensor(image_grid_thw)
-        else:
-            raise ValueError("No image grid provided.")
 
-        return {
-            "image": {
-                "image_embeds": image_embeds.squeeze(0),
-                "image_grid_thw": grid_thw_tensor,
-            }
-        }
-    elif "llava-hf/LLaVA-NeXT-Video-7B-hf" in model:
+    # Handle video models
+    if model == SupportedModels.LLAVA_NEXT_VIDEO_7B:
         if video_numpy is None:
             raise ValueError("No video frames provided.")
         return {"video": video_numpy}
+
+    # Handle image models - validate image embeddings first
+    if image_embeds is None:
+        raise ValueError("No image embeddings provided.")
+
+    image_embeds = image_embeds.to(embeddings_dtype)
+
+    # Model-specific image handling
+    if model == SupportedModels.QWEN_2_5_VL_7B:
+        return _construct_qwen_image_data(image_embeds, image_grid_thw)
     else:
-        if image_embeds is None:
-            raise ValueError("No image embeddings provided.")
-        image_embeds = image_embeds.to(embeddings_dtype)
+        # Default image handling for other models (e.g., LLAVA_1_5_7B)
         return {"image": image_embeds}
+
+
+def _construct_qwen_image_data(
+    image_embeds: torch.Tensor, image_grid_thw: Optional[List[Any]]
+) -> Dict[str, Dict[str, torch.Tensor]]:
+    """Construct image data specifically for Qwen models."""
+    if image_grid_thw is None or len(image_grid_thw) == 0:
+        raise ValueError("No image grid provided for Qwen model.")
+
+    grid_thw_tensor = torch.tensor(image_grid_thw)
+
+    return {
+        "image": {
+            "image_embeds": image_embeds.squeeze(0),
+            "image_grid_thw": grid_thw_tensor,
+        }
+    }
